Added graphing abilities to output display. finished the tree. given andreea shit loads of work to do. taken more caffeine than is sane/safe.

+++ gwt/ReefFront/src/uk/ac/imperial/vazels/reef/client/output/TimeStamp.java
@@ -10,8 +10,8 @@ public class TimeStamp extends JavaScriptObject {
     return this.type;
   }-*/;
   
-  native final int getDouble() /*-{
-    return parseInt(this.value);
+  native final float getDouble() /*-{
+    return parseFloat(this.value);
   }-*/;
   
   native final String getString() /*-{
